+++ src-tauri/src/ipc/mod.rs
@@ -1,4 +1,5 @@
 pub mod cmd_audio;
+pub mod cmd_button_color;
 pub mod cmd_button_flags;
 pub mod cmd_button_playback;
 pub mod cmd_button_types;